+++ .gitignore
@@ -1,1 +1,2 @@
 score.json
+__pycache__/
